Force float32 for bias_AAs_np for MPS compatibility
bias_AAs_np was created via np.zeros(len(alphabet)) which defaults to
float64. When passed to torch.tensor() with device=mps inside
ProteinMPNN.sample (protein_mpnn_utils.py:1135), this triggers:

  TypeError: Cannot convert a MPS Tensor to float64 dtype as the MPS
  framework doesn't support float64. Please use float32 instead.

Fix at the source by allocating the array as float32, matching the
existing float32 convention for omit_AAs_np a few lines above. CPU
and CUDA paths are unaffected by the explicit dtype.

Discovered while running test_mps_e2e.py end-to-end on 6MRR.

+++ protein_mpnn_run.py
@@ -164,7 +164,7 @@ def main(args):
 
     if print_all: 
         print(40*'-')
-    bias_AAs_np = np.zeros(len(alphabet))
+    bias_AAs_np = np.zeros(len(alphabet), dtype=np.float32)
     if bias_AA_dict:
             for n, AA in enumerate(alphabet):
                     if AA in list(bias_AA_dict.keys()):
